+++ src/cmd/browse/layout.rs
@@ -27,7 +27,7 @@ use crate::cmd::browse::entries::entry_schedule;
 use crate::cmd::browse::entries::entry_type_label;
 use crate::cmd::browse::entries::entry_url;
 use crate::cmd::browse::state::BrowseState;
-use crate::cmd::browse::template::page_template;
+use crate::cmd::browse::templates::page_template;
 use crate::error::Fallible;
 use crate::types::aliases::DeckName;
