+++ src/cmd/browse/views/card_basic.rs
@@ -26,10 +26,10 @@ use crate::cmd::browse::render::performance_rows;
 use crate::cmd::browse::render::render_config;
 use crate::cmd::browse::render::render_history;
 use crate::cmd::browse::render::source_rows;
+use crate::cmd::browse::shared::error_response;
+use crate::cmd::browse::shared::internal_error_response;
+use crate::cmd::browse::shared::ok_response;
 use crate::cmd::browse::state::BrowseState;
-use crate::cmd::browse::template::error_response;
-use crate::cmd::browse::template::internal_error_response;
-use crate::cmd::browse::template::ok_response;
 use crate::error::Fallible;
 use crate::types::card::Card;
 use crate::types::card_hash::CardHash;
